requirements.txt
accelerate==1.0.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.4.0
diffusers==0.31.0
distro==1.9.0
filelock==3.16.1
fsspec==2024.10.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.26.1
idna==3.10
importlib_metadata==8.5.0
inquirerpy==0.3.4
Jinja2==3.1.4
lxml==5.3.0
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==1.26.4
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.3
pfzy==0.3.4
pillow==11.0.0
praw==7.8.1
prawcore==2.4.0
prompt_toolkit==3.0.48
protobuf==5.28.3
psutil==6.1.0
pydantic==2.9.2
pydantic_core==2.23.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
safetensors==0.4.5
sentencepiece==0.2.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sympy==1.13.3
tokenizers==0.20.1
torch==2.2.2
tqdm==4.66.5
transformers==4.46.0
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
update-checker==0.18.0
urllib3==2.2.3
wcwidth==0.2.13
websocket-client==1.8.0
wsproto==1.2.0
zipp==3.20.2
